+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New method `changeName()` that does the same as the native `rename()` function, and will throw in case the file could not be renamed.
+- New function `writeOnFile()` that does the same as the native `fwrite()` function, and will throw in case the file could not be written.
 
 ## [0.1.0] 2020-09-11